refactor(courses): Use a single file for all domain errors
