--- src/courses/domain/errors/no_course_with_invitation_code_error.go
@@ -1,18 +1,0 @@
-package errors
-
-import (
-	"fmt"
-	"net/http"
-)
-
-type NoCourseWithInvitationCodeError struct {
-	Code string
-}
-
-func (err NoCourseWithInvitationCodeError) Error() string {
-	return fmt.Sprintf("Course with invitation code %s not found", err.Code)
-}
-
-func (err NoCourseWithInvitationCodeError) StatusCode() int {
-	return http.StatusNotFound
-}
