--- src/courses/domain/errors/teacher_does_not_owns_course_error.go
@@ -1,16 +1,0 @@
-package errors
-
-import (
-	"net/http"
-)
-
-type TeacherDoesNotOwnsCourseError struct {
-}
-
-func (err TeacherDoesNotOwnsCourseError) Error() string {
-	return "You do not own the course"
-}
-
-func (err TeacherDoesNotOwnsCourseError) StatusCode() int {
-	return http.StatusForbidden
-}
