--- src/courses/domain/errors/user_not_in_course_error.go
@@ -1,15 +1,0 @@
-package errors
-
-import (
-	"net/http"
-)
-
-type UserNotInCourseError struct{}
-
-func (err UserNotInCourseError) Error() string {
-	return "You are not enrolled in the course"
-}
-
-func (err UserNotInCourseError) StatusCode() int {
-	return http.StatusForbidden
-}
